+++ lib/core/app_assets.dart
@@ -1,4 +1,5 @@
 abstract class AppAssets {
   static const _svgPath = 'assets/svgs';
-  static const mainIconNoBackground = '$_svgPath/main-icon-no-background.svg';
+  static const mainIcon = '$_svgPath/main_icon.svg';
+  static const mainIconLong = '$_svgPath/main_icon_long.svg';
 }