+++ lib/core/widgets/cards/app_card.dart
@@ -1,4 +1,5 @@
 import 'package:awesome_period_tracker/core/extensions/build_context_extensions.dart';
+import 'package:awesome_period_tracker/core/extensions/color_extensions.dart';
 import 'package:flutter/material.dart';
 
 class AppCard extends StatelessWidget {
@@ -21,12 +22,16 @@ class AppCard extends StatelessWidget {
       decoration: BoxDecoration(
         color: backgroundColor ?? context.colorScheme.surface,
         borderRadius: BorderRadius.circular(16),
-        border: border ??
-            Border.all(
-              color: context.colorScheme.shadow.withAlpha(30),
-              width: 1,
-            ),
-        boxShadow: boxShadow,
+        border: border,
+        boxShadow: boxShadow ??
+            [
+              BoxShadow(
+                color:
+                    context.colorScheme.primary.darken(0.6).withOpacity(0.15),
+                blurRadius: 8,
+                offset: const Offset(0, 2),
+              ),
+            ],
       ),
       child: child,
     );